[Legacy line layout removal] Remove special empty inline handling https://bugs.webkit.org/show_bug.cgi?id=271818 rdar://problem/125542223
Reviewed by NOBODY (OOPS!).

In SVG whitespaces always collapse.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/text/BidiResolver.h:
(WebCore::WhitespaceCollapsingState::ensureLineBoxInsideIgnoredSpaces): Deleted.
* Source/WebCore/rendering/LegacyLineLayout.h:
* Source/WebCore/rendering/RenderBlockFlow.h:
* Source/WebCore/rendering/line/BreakingContext.h:
(WebCore::updateWhitespaceCollapsingTransitionsForTrailingWhitespace):
(WebCore::BreakingContext::handleText):
(WebCore::BreakingContext::commitAndUpdateLineBreakIfNeeded):
(WebCore::BreakingContext::handleEndOfLine):
(WebCore::shouldSkipWhitespaceAfterStartObject): Deleted.
(WebCore::BreakingContext::handleEmptyInline): Deleted.
(WebCore::checkWhitespaceCollapsingTransitions): Deleted.
* Source/WebCore/rendering/line/LineBreaker.cpp:
(WebCore::LineBreaker::nextLineBreak):
* Source/WebCore/rendering/line/TrailingObjects.cpp: Removed.
* Source/WebCore/rendering/line/TrailingObjects.h: Removed.

Replace with a field in BreakingContext.
